+++ test/integrated_tests/core/authc/test_authentication.py
@@ -2,13 +2,14 @@ import pytest
 from yosai.core import (
     AuthenticationException,
     Credential,
+    InvalidAuthenticationSequenceException,
     SerializationManager,
     UsernamePasswordToken,
     AuthenticationException,
 )
 
 
-def test_authentication_using_accountstore_success(
+def test_single_factor_authc_userpass_using_accountstore_success(
         caplog, default_authenticator, valid_username_password_token,
         event_bus):
     da = default_authenticator
@@ -19,13 +20,39 @@ def test_authentication_using_accountstore_success(
         event_detected = identifiers
     event_bus.register(event_listener, 'AUTHENTICATION.SUCCEEDED')
 
-    account = da.authenticate_account(valid_username_password_token)
+    account = da.authenticate_account(None, valid_username_password_token)
     out = caplog.text
     assert (event_detected == account.account_id and
             ("Could not obtain cached" in out and "No account" not in out))
 
 
-def test_authentication_using_cache_success(
+def test_multi_factor_authc_raises_invalid_sequence(
+        default_authenticator, valid_thedude_totp_token, monkeypatch):
+    da = default_authenticator
+
+    with pytest.raises(InvalidAuthenticationSequenceException):
+        da.authenticate_account(None, valid_thedude_totp_token)
+
+
+def test_multi_factor_authc_using_accountstore_success(
+        caplog, default_authenticator, valid_username_password_token,
+        valid_totp_token, event_bus):
+    da = default_authenticator
+    event_detected = None
+
+    def event_listener(identifiers=None):
+        nonlocal event_detected
+        event_detected = identifiers
+    event_bus.register(event_listener, 'AUTHENTICATION.SUCCEEDED')
+
+    account = da.authenticate_account(None, valid_username_password_token)
+    out = caplog.text
+    assert (event_detected == account.account_id and
+            ("Could not obtain cached" in out and "No account" not in out))
+
+
+
+def test_single_factor_authc_userpass_using_cache_success(
         caplog, default_authenticator, invalid_username_password_token,
         valid_username_password_token, cache_handler, event_bus):
 
@@ -38,9 +65,9 @@ def test_authentication_using_cache_success(
     event_bus.register(event_listener, 'AUTHENTICATION.SUCCEEDED')
 
     with pytest.raises(AuthenticationException):
-        da.authenticate_account(invalid_username_password_token)
+        da.authenticate_account(None, invalid_username_password_token)
 
-        account = da.authenticate_account(valid_username_password_token)
+        account = da.authenticate_account(None, valid_username_password_token)
 
         out = caplog.text
         assert (event_detected == account.account_id and
@@ -48,7 +75,7 @@ def test_authentication_using_cache_success(
                 account.account_id == valid_username_password_token.identifier)
 
 
-def test_authentication_using_accountstore_pw_failure(
+def test_single_factor_authc_userpass_using_accountstore_failure(
         caplog, default_authenticator, invalid_username_password_token,
         event_bus):
 
@@ -61,14 +88,14 @@ def test_authentication_using_accountstore_pw_failure(
     event_bus.register(event_listener, 'AUTHENTICATION.FAILED')
 
     with pytest.raises(AuthenticationException):
-        account = da.authenticate_account(invalid_username_password_token)
+        account = da.authenticate_account(None, invalid_username_password_token)
 
         out = caplog.text
         assert (event_detected == account.account_id and
                 ("Could not obtain cached" in out and "No account" not in out))
 
 
-def test_authentication_using_cache_pw_failure(
+def test_single_factor_authc_userpass_using_cache_failure(
         caplog, default_authenticator, invalid_username_password_token,
         cache_handler, event_bus):
 
@@ -84,7 +111,7 @@ def test_authentication_using_cache_pw_failure(
     cache_handler.set(domain='credentials', identifier='thedude', value=cred)
 
     with pytest.raises(AuthenticationException):
-        da.authenticate_account(invalid_username_password_token)
+        da.authenticate_account(None, invalid_username_password_token)
         out = caplog.text
 
         assert (event_detected == invalid_username_password_token.username and
@@ -92,7 +119,7 @@ def test_authentication_using_cache_pw_failure(
 
     cache_handler.delete(domain='credentials', identifier='thedude')
 
-def test_authentication_using_accountstore_user_not_found(
+def test_single_factor_authc_userpass_using_accountstore_user_not_found(
         default_authenticator, event_bus):
     da = default_authenticator
     event_detected = None
@@ -108,6 +135,6 @@ def test_authentication_using_accountstore_user_not_found(
                                        host='127.0.0.1')
 
     with pytest.raises(AuthenticationException):
-        da.authenticate_account(dumb_token)
+        da.authenticate_account(None, dumb_token)
 
     assert (event_detected == dumb_token.username)