+++ yosai/core/account/abcs.py
@@ -18,100 +18,10 @@ under the License.
 """
 
 from abc import ABCMeta, abstractmethod
-from yosai.core import (
-    serialize_abcs,
-)
-
-
-# not pythonic, will change later:
-class AccountId(metaclass=ABCMeta):
-    @abstractmethod
-    def __repr__(self):
-        pass
-
-
-class Account(metaclass=ABCMeta):
-    """
-    An Account is a unique identity within an AccountStore  that has a set of
-    attributes.  An account may represent a human being, but this is not
-    required - an account could represent a host, a server, a daemon -
-    basically anything with an identity that might need to be authenticated or
-    authorized to perform behavior.
-
-    Implementation Warning
-    ----------------------
-    Since Yosai sometimes logs account operations, please ensure your Account's
-    __repr__ implementation does *not* print out account CREDENTIALS
-    (password, etc), as these might be viewable to someone reading your logs.
-    This is good practice anyway, and account identifiers should rarely (if
-    ever) be printed out for any reason.
-
-    Yosai's default implementations of this interface only ever print account
-    attributes.
-
-    """
-
-    @property
-    @abstractmethod
-    def account_id(self):  # DG:  renamed
-        """
-        Returns an identifier unique compared to any other Account found in the
-        same account store.  For example, this can be a store-wide unique
-        username or email address, database primary key, UUID, GUID, etc.
-
-        After an account is authenticated, Yosai will use this id for all
-        future access to the account: for caching the account (if caching is
-        enabled), for future lookups from the account store, and any lookups
-        for authorization.
-
-        :returns: an identifier unique compared to any other Account found in
-                  the same account store
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def credentials(self):
-        """
-        Returns the stored credentials associated with the corresponding
-        Account.  Credentials, such as a password or private key, verifies one
-        or more submitted identities during authentication.
-
-        Shiro references credentials during the authentication process to
-        ensure that submitted credentials during a login attempt match exactly
-        the Account's stored credentials returned via this method.
-
-        :returns: the credentials associated with the corresponding Subject
-        """
-        pass
-
-    # yosai.core.renamed "attributes" to identifiers:
-    @property
-    @abstractmethod
-    def attributes(self):
-        """
-        Returns an *immutable* view of the Account's attributes,
-        excluding the AccountID, accessible to the application.  Once the
-        account is obtained, an application developer can access them as
-        desired, for example:
-
-            username = account.attributes.get('username')
-            print("Welcome, " + username + "!")
-
-        :returns: the Account's attributes accessible to the application
-
-        Requires a corresponding AccountAttributesSchema class that defines its
-        serialization schema.
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def authz_info(self):
-        pass
 
 
 class AccountStore(metaclass=ABCMeta):
+    pass
 
 #    @abstractmethod
 #    def get_account(self, request):
@@ -127,6 +37,7 @@ class AccountStore(metaclass=ABCMeta):
 #        :returns: Account
 #        """
 #        pass
+
 
 class LockingAccountStore(AccountStore):
 
@@ -155,17 +66,3 @@ class AuthorizationAccountStore(AccountStore):
         :returns: Account
         """
         pass
-
-    # @abstractmethod
-    # def get_permissions(self, identifiers):
-    #    """
-    #    :returns: Account
-    #    """
-    #    pass
-
-    # @abstractmethod
-    # def get_roles(self, identifiers):
-    #    """
-    #    :returns: Account
-    #    """
-    #    pass