+++ yosai/core/account/account.py
@@ -1,52 +1,8 @@
-from yosai.core import (
-    account_abcs,
-)
+from collections import namedtuple
 
-
-class Account(account_abcs.Account):
-    """
-    This is a basic collection of the security related attributes associated
-    with an account
-    """
-
-    def __init__(self, account_id=None, credentials=None, attributes=None,
-                 authz_info=None):
-        self.account_id = account_id
-        self.credentials = credentials
-        self.attributes = attributes
-        self.authz_info = authz_info
-
-    @property
-    def account_id(self):
-        return self._account_id
-
-    @account_id.setter
-    def account_id(self, accountid):
-        self._account_id = accountid
-
-    @property
-    def credentials(self):
-        return self._credentials
-
-    @credentials.setter
-    def credentials(self, credentials):
-        self._credentials = credentials
-
-    @property
-    def attributes(self):
-        return self._attributes
-
-    @attributes.setter
-    def attributes(self, attributes):
-        self._attributes = attributes
-
-    @property
-    def authz_info(self):
-        return self._authz_info
-
-    @authz_info.setter
-    def authz_info(self, authz_info):
-        self._authz_info = authz_info
-
-    def __repr__(self):
-        return "Account(account_id={0})".format(self._account_id)
+# this is a namedtuple with default values, it may not get used (TBD)
+class Account(namedtuple('Account', 'account_id, account_locked, authc_info, authz_info')):
+    __slots__ = ()
+    def __new__(cls, account_id, account_locked=None, authc_info=None, authz_info=None):
+        return super(Account, cls).__new__(cls, account_id, account_locked,
+                                           authc_info, authz_info)