+++ yosai/web/mgt/mgt.py
@@ -165,13 +165,13 @@ class WebSecurityManager(NativeSecurityManager):
             pass
 
     # new to yosai, overriding to support CSRF token synchronization
-    def on_successful_login(self, authc_token, account, subject):
+    def on_successful_login(self, authc_token, account_id, subject):
 
         # Generating a new session_id at successful login is a recommended
         # countermeasure to a session fixation
         subject.session = subject.session.recreate_session()
 
-        super().remember_me_successful_login(authc_token, account, subject)
+        super().remember_me_successful_login(authc_token, account_id, subject)
 
 
 class CookieRememberMeManager(AbstractRememberMeManager):